Refactor CommandLineTool decorator into its own module and update imports in main.py
